Fix up test cases to reflect moved methods
